requirements.txt
python-dotenv==0.19.2
tabulate==0.8.9
python-telegram-bot==13.10